fix: 변수 type 변경 (#65)
jwt가 서버로 돌아올 때는 Long을 유지하지 못하므로 Integer로 통일

+++ src/main/java/com/newfit/reservation/common/jwt/TokenProvider.java
@@ -67,8 +67,8 @@ public class TokenProvider {
     }
 
     private void checkAuthorityIdList(String token, HttpServletRequest request) {
-        List<Long> authorityIdList = getAuthorityIdList(token);
-        Long authorityId = authorityRepository.findOne(Long.parseLong(request.getHeader("authorityId"))).orElseThrow(IllegalArgumentException::new).getId();
+        List<Integer> authorityIdList = getAuthorityIdList(token);
+        Integer authorityId = authorityRepository.findOne(Long.parseLong(request.getHeader("authorityId"))).orElseThrow(IllegalArgumentException::new).getId().intValue();
         authorityIdList
                 .stream()
                 .filter(authority -> authority.equals(authorityId))
